backend/requirements.txt
Django==6.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.10
dj-database-url==3.0.1
python-decouple==3.8
Pillow==12.1.0
openpyxl==3.1.5
requests==2.32.3
asgiref==3.11.0
PyJWT==2.10.1
sqlparse==0.5.5
whitenoise==6.9.0
gunicorn==23.0.0
